+++ R/adv.R
@@ -1135,7 +1135,7 @@ setMethod(f="plot",
                   }
               }
               oceDebug(debug, "} # plot.adv()\n", unindent=1)
-              invisible()
+              invisible(NULL)
           })