+++ create_data/section/create_section.R
@@ -2,7 +2,7 @@ library(oce)
 ## next worked 2017-02-20
 ## download.file("https://cchdo.ucsd.edu/data/7872/a03_hy1.csv", "a03_hy1.csv")
 section <- read.section("a03_hy1.csv", sectionId="a03", institute="SIO",
-                        ship="R/V Professor Multanovskiy", scientist="Vladimir Tereschenkov") 
+                        ship="R/V Professor Multanovskiy", scientist="Vladimir Tereschenkov")
 section <- initializeFlagScheme(section, "WHP bottle")
 
 ## Save in version 2, because otherwise users with R 3.5.x and earlier will not